+++ examples/example-axum/src/extractors.rs
@@ -1,6 +1,6 @@
 use aide::operation::OperationIo;
 use axum::response::IntoResponse;
-use axum_jsonschema::JsonSchemaRejection;
+use axum_jsonschema::extract::{JsonRejection, SerdeSchemaRejection};
 use axum_macros::FromRequest;
 use serde::Serialize;
 use serde_json::json;
@@ -8,10 +8,10 @@ use serde_json::json;
 use crate::errors::AppError;
 
 #[derive(FromRequest, OperationIo)]
-#[from_request(via(axum_jsonschema::Json), rejection(AppError))]
+#[from_request(via(axum_jsonschema::extract::Json), rejection(AppError))]
 #[aide(
-    input_with = "axum_jsonschema::Json<T>",
-    output_with = "axum_jsonschema::Json<T>",
+    input_with = "axum_jsonschema::extract::Json<T>",
+    output_with = "axum_jsonschema::extract::Json<T>",
     json_schema
 )]
 pub struct Json<T>(pub T);
@@ -25,15 +25,16 @@ where
     }
 }
 
-impl From<JsonSchemaRejection> for AppError {
-    fn from(rejection: JsonSchemaRejection) -> Self {
+impl From<JsonRejection> for AppError {
+    fn from(rejection: JsonRejection) -> Self {
         match rejection {
-            JsonSchemaRejection::Json(j) => Self::new(&j.to_string()),
-            JsonSchemaRejection::Path(_) => unreachable!("Not using Path extractor"),
-            JsonSchemaRejection::Serde(_) => Self::new("invalid request"),
-            JsonSchemaRejection::Schema(s) => {
-                Self::new("invalid request").with_details(json!({ "schema_validation": s }))
-            }
+            JsonRejection::Json(j) => Self::new(&j.to_string()),
+            JsonRejection::SerdeSchema(e) => match e {
+                SerdeSchemaRejection::Serde(_) => Self::new("invalid request"),
+                SerdeSchemaRejection::Schema(s) => {
+                    Self::new("invalid request").with_details(json!({ "schema_validation": s }))
+                }
+            },
         }
     }
 }
